Script Migration and Module Testing Migrated most scripts, and tested Client Modules for correctness.
Increased WebSocket TImeout for SignalR, should help for connection stability till .NET 5 RC1.

Modules
- Details
- Interaction
- Interaction Indicator
- Stopping, Finished implementation

Client Actions
- MessageFromCombatSystem
- RunSkillAction

Added Localization of Templates with Key/Value.
Simplified ObjectEetityDetailsModel, will now default Id.
Added better logging to LifecycleEntityBase.
Updated Actions namespaces to ClientAction namespace.

Code Cleanup
- Warnings
- Formatting
- MaybeNull
- Updated to use IsNull/IsNotNull
- Usage of Option<>, instead of allowing null

9EC1F9B3-6806-411F-8704-C858680D5D32

+++ src/EventHorizon.Game.Client.Systems/Player/Modules/PlayerInteraction/Model/StandardPlayerInteractionModule.cs
@@ -1,9 +1,7 @@
 ﻿namespace EventHorizon.Game.Client.Systems.Player.Modules.PlayerInteraction.Model
 {
     using System;
-    using System.Collections;
     using System.Collections.Generic;
-    using System.Collections.Immutable;
     using System.Linq;
     using System.Threading.Tasks;
     using EventHorizon.Game.Client.Engine.Systems.ClientAction.Publish;
@@ -99,7 +97,6 @@
                 .Select(a => a.Value);
             if (interactionItemList.Any())
             {
-
                 await _mediator.Publish(
                     new ShowInteractionIndicatorEvent(
                         interactionItemList
@@ -117,11 +114,11 @@
         private struct InteractionItem
         {
             public IObjectEntity Entity { get; }
-            public float DistanceToPlayer { get; }
+            public decimal DistanceToPlayer { get; }
 
             public InteractionItem(
                 IObjectEntity entity,
-                float distanceToPlayer
+                decimal distanceToPlayer
             )
             {
                 Entity = entity;
